# optimize queries for courses list page