[**EDUCATOR-828**](https://openedx.atlassian.net/browse/EDUCATOR-828)

These changes will improve the publisher courses list page(`/publisher/courses`) load time. There is still room for improvement but for now this good enough.

-----------
**Publisher courses page load time with queries executed for `1242` courses** 

Before Optimisation
——————————
1371.67 ms (8875 queries including 8869 duplicates )

After Optimisation
—————————
561.99 ms (3912 queries including 3901 duplicates )

-----------

**Unit test stats shows queries executed before and after the optimization**

|               test                                            | before    |  after|
----------------------------------------|----------|-------|
|test_page_with_enable_waffle_switch    |     13         |   15|
|test_page_with_disable_waffle_switch   |      21        |    21|
|test_courses_with_permission               |    100         |  64|
|test_courses_with_course_user_role      |    60          |  32|
|test_courses_with_admin                        |     58         |   31 |