@MichaelRoytman 2017-07-14
Hello,

I saw that this commit went to stage this morning. I got a bunch of pages this morning for stage discovery.

You can see the NewRelic alerts below.

https://rpm.newrelic.com/accounts/88178/incidents/30400026

https://rpm.newrelic.com/accounts/88178/applications/15120497/traced_errors/8a5f2e9e-68a4-11e7-8968-0242ac11000f_16360_21050/similar_errors?original_error_id=8a5f2e9e-68a4-11e7-8968-0242ac11000f_16360_21050

Could you take a look at this before it goes to production? Please coordinate with the Learner - Spartan team, if need be.

Best,
Michael