@muzaffaryousaf 2017-07-14
@MichaelRoytman I've verified from New Relic https://rpm.newrelic.com/accounts/88178/applications/15120497/traced_errors/8a5f2e9e-68a4-11e7-8968-0242ac11000f_16360_21050?original_error_id=8a5f2e9e-68a4-11e7-8968-0242ac11000f_16360_21050
This PR did't changed anything related to this error, something else must be. Looks like something is wrong with Publisher's data.